+++ src/Saga/Utils/TypeLevel.hs
@@ -1,0 +1,4 @@
+module Saga.Utils.TypeLevel where
+
+
+type a $ b = a b
